P2 server: decompose WebSocketHost into Hosting/ components
Split the ~550-line god class into focused, individually testable pieces
under src/B3.Umdf.Server/Hosting/:

- HealthEvaluator           pure stale-recovery rule for /health (no IO)
- HealthEndpointMapper      /health, /ready, /live wiring
- MetricsEndpointMapper     OTel meter registration + /metrics scrape
- SymbolEndpointMapper      /symbols
- InstrumentEndpointMapper  /instrument/{symbol} + DTO mapping
- RecoveryEndpointMapper    /api/recovery/recent
- WebSocketConnectionHandler  /ws lifecycle (accept, hello, sub/unsub, drain)
- ShutdownCoordinator       graceful 1001 close-handshake fan-out
- CorsMiddleware            permissive CORS extension
- ClientSessionOptions      grouped per-connection tuning record

WebSocketHost retains its public ctor + property surface and now only
orchestrates: build the Kestrel app, hand each concern to its mapper,
and expose the providers wired in by the ConsoleApp. No behavior change;
all 150 existing Server tests pass unmodified.
